Add scenarios based on Bustabit data
This scenario is composed from two data sources:

• Bustabit's payment data via
https://github.com/Xekyo/CoinSelectionSimulator/commit/e3ebe0209c1f17cd2358462d6592fb3641f655d6
• 6-block target fee rate estimates from Bitcoin Core with one data
point per day over the period of 2019–2020 via txstats.com

The payments are sourced from Bustabit's (a service) deposits and
withdrawals. The sequence of payments has about twice as many deposits
as withdrawals, with a total of about 540k incoming/outgoing payments in
sum. The payment sequence was zipped with the about 730 daily fee rate estimates from
2019-2020 extracted via txstats.com. The sequence of fee rate estimates
ramps up from low fee rates to a peak, goes back into a lull and ramps
up to a second period of increased feerates. The fee rate pattern simply
repeats to match the length of the much longer payment sequence.

The `bustabit-2019-2020-short.csv` abbreviates this to only the first
50k payments, which includes about 16k withdrawals.